Fix issue with sample size being an odd number
Current problem related to Issue #4 was actually that the method didn't work for any odd numbers, instead of just small numbers.

Method is based on trying to cover a triangle matrix but using sample_size//2 meant the method would miss meeting the main diagonal due to loss of prevision of the floor function (the method assumes it will reach the main diagonal to allow us to take better sets), causing sets taken near to the main diagonal to be too small and others to miss areas of the matrix. 

To fix it if sample_size is odd I solve for sample_size -1 instead and then fix the sample size.